+++ source/components/resizable-container/index.js
@@ -1,0 +1,2 @@
+import "./resizable-container.less";
+import "./resizable-container.js";