+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 + [x] Fix validator for the `required` rule for the `input` with role
 + [x] Improved rule `date` for validator
++ [x] Added component `resizable-container`
++ [x] Fixed bug in `hotkey` component when using `$(...).hotkey()` method